Refactor role.query.ts and add useUserRoleCreate, useUserRoleList, useUserRoleEdit, useUserRoleDelete hooks
